requirements.txt
boto3==1.42.25
dj-database-url==3.1.0
django-storages==1.14.6
django-webpack-loader==3.2.3
gunicorn==23.0.0
pillow_heif==1.1.1
pip==25.3
psycopg2-binary==2.9.11
pytest==9.0.2
sentry-sdk==2.49.0
setuptools==80.9.0
wagtail==7.2.1
whitenoise==6.11.0
mistralai==1.10.0
djangorestframework==3.16.1
pypandoc==1.17
weasyprint==68.1
